Fix issue #1 - Skip auth only if the current principal is for the expected SP.
I'm not sure this is the right way - Maybe the idea here is to avoid going back to the IdP for each new service provider, in which case we should perhaps look to change how the principal is used to avoid it 'locking-in' to the first service provider used. It's also entirely possible I've misconfigured or misunderstood something about how the proxy is supposed to work - Pointers much appreciated if so!